Added code to check

+++ src/main/java/com/example/resttemplateforquotersextendedapplication/QuoteValue.java
@@ -1,4 +1,4 @@
 package com.example.resttemplateforquotersextendedapplication;
 
-public record QuoteValue(Integer id, String quote) {
+public record QuoteValue(Long id, String quote) {
 }
